# server/docs/swagger.yaml
basePath: /
definitions:
  common.ErrorResponseDto:
    properties:
      error:
        type: string
    type: object
  image.CreateImageRequest:
    properties:
      url:
        type: string
    type: object
  image.GetImageResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user:
        type: string
    type: object
  image.UpdateImageRequest:
    properties:
      url:
        type: string
    type: object
  presign.PresignedObjectRequest:
    properties:
      originalName:
        type: string
      type:
        type: string
    type: object
  presign.PresignedObjectResponse:
    properties:
      bucketName:
        type: string
      key:
        type: string
      url:
        type: string
    type: object
  user.CreateUserRequest:
    properties:
      name:
        maxLength: 50
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - name
    - password
    - username
    type: object
  user.GetUserResponse:
    properties:
      id:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  user.UpdateUserRequest:
    properties:
      name:
        maxLength: 50
        minLength: 3
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API for MyPic application
  title: MyPic API
  version: "1.0"
paths:
  /api/v1/image:
    get:
      consumes:
      - application/json
      description: Get all images in the database
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: number
      - default: 10
        description: Number of images per page
        in: query
        name: limit
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/image.GetImageResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
      summary: Get all images
      tags:
      - Images
    post:
      consumes:
      - application/json
      description: Create a new image with the provided details
      parameters:
      - description: Image details
        in: body
        name: image
        required: true
        schema:
          $ref: '#/definitions/image.CreateImageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/image.GetImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
      summary: Create a new image
      tags:
      - Images
  /api/v1/image/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing image by ID
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
      summary: Delete an existing image
      tags:
      - Images
    get:
      consumes:
      - application/json
      description: Get an image by its unique ID
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/image.GetImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
      summary: Get an image by ID
      tags:
      - Images
    put:
      consumes:
      - application/json
      description: Update an existing image with the provided details
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated image details
        in: body
        name: image
        required: true
        schema:
          $ref: '#/definitions/image.UpdateImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/image.GetImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
      summary: Update an existing image
      tags:
      - Images
  /api/v1/presign:
    post:
      consumes:
      - application/json
      description: Get presigned URL for uploading an object to S3
      parameters:
      - description: Presigned Object Request
        in: body
        name: presignedObjectRequest
        required: true
        schema:
          $ref: '#/definitions/presign.PresignedObjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presign.PresignedObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
      summary: Get presigned URL for uploading an object to S3
      tags:
      - Presign
  /api/v1/user:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided details
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
      summary: Create a new user
      tags:
      - Users
  /api/v1/user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
      summary: Delete an existing user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get a user by their unique ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
      summary: Get a user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update an existing user with the provided details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated user details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponseDto'
      summary: Update an existing user
      tags:
      - Users
swagger: "2.0"